fix!: fix add Test to the name of the classes so jUnit detects it
